Approval on update and delete added on books categories

+++ app/Http/Controllers/BooksCategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\BooksCategory;
+use App\Models\ModificationRequest;
 use Illuminate\Support\Facades\Auth;
 
 class BooksCategoryController extends Controller
@@ -15,7 +16,7 @@ class BooksCategoryController extends Controller
      */
     public function index()
     {
-        $booksCategories = BooksCategory::withCount('books')->paginate(10);
+        $booksCategories = BooksCategory::withCount('books')->orderBy('id', 'desc')->paginate(10);
 
         return view('booksCategory.index', compact('booksCategories'));
     }
@@ -74,12 +75,30 @@ class BooksCategoryController extends Controller
         ]);
 
         $book = BooksCategory::findOrFail($id);
-        $book->category_name = $request->category_name;
-        $book->published_volumns = $request->published_volumns;
-        $book->published_total_volumns = $request->published_total_volumns;
-        $book->save();
+        if ($book) {
+            if (auth()->user()->hasRole('president')) {
 
-        return redirect()->route('bookCategories')->with('success', 'Books category updated successfully.');
+                $book->category_name = $request->category_name;
+                $book->published_volumes = $request->published_volumes;
+                $book->published_total_volumes = $request->published_total_volumes;
+
+                $book->save();
+                return redirect()->route('bookCategories')->with('success', 'Books category updated successfully.');
+            } else {
+
+                $changes = $request->except(['_token', '_method']);
+                $this->submitChangeRequest([
+                    "table_name" => 'books_categories',
+                    "record_id" => $book->id,
+                    "changes" => $changes,
+                    "action" => ModificationRequest::REQUEST_TYPE_UPDATE,
+                    "requested_by" => Auth::id(),
+                ]);
+            }
+            return redirect()->route('bookCategories')->with('success', 'Books category update request submitted successfully.');
+        }
+
+        return redirect()->route('bookCategories')->with('error', 'Something went wrong.');
     }
 
     public function destroy($id)
@@ -87,12 +106,25 @@ class BooksCategoryController extends Controller
         // Find the user by ID
         $book = BooksCategory::findOrFail($id);
 
-        $book->deleted_by = Auth::id();
-        $book->save();
+        if ($book) {
+            if (auth()->user()->hasRole('president')) {
+                $book->deleted_by = Auth::id();
+                $book->save();
 
-        // Soft delete the book
-        $book->delete();
+                // Soft delete the book
+                $book->delete();
 
-        return redirect()->route('bookCategories')->with('success', 'Book category deleted successfully!');
+                return redirect()->route('bookCategories')->with('success', 'Book category deleted successfully!');
+            } else {
+                $this->submitChangeRequest([
+                    "table_name" => 'books_categories',
+                    "record_id" => $book->id,
+                    "action" => ModificationRequest::REQUEST_TYPE_DELETE,
+                    "requested_by" => Auth::id(),
+                ]);
+                return redirect()->route('locations')->with('success', 'Location delete request submitted successfully!');
+            }
+            return redirect()->route('bookCategories')->with('error', 'Something went wrong.');
+        }
     }
 }
